compartmentalized functionality
